Added checks for ownership and existence
Issue 874: Squashed commit of issue-874b branch;

    commit c6ff7037b381ef488358564cb3f829f36d5ae09a
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 16:13:05 2022 -0400

        Added a check to make sure a file exists and is not zero length before attempting to transfer it

    commit 45867bb0d7e0196d74578903cd48b091422c8bd5
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 16:01:25 2022 -0400

        Added break to avoid infinite loop

    commit 406b30ff2bb68b133d7470761032becd727378c1
    Merge: a1f6f17 e568254
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 15:52:02 2022 -0400

        Merge branch 'issue-874b' of https://github.com/StormSurgeLive/asgs into issue-874b

    commit a1f6f176ce55dafecdf352848f1cab2cf2cba314
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 15:51:57 2022 -0400

        Comparing remote user and owner locally

    commit e5682542c680ed32169a6052f63c7e22d4793d88
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 13:59:29 2022 -0500

        Checked to make sure asgs instance status file existed before attempting to post it.

    commit b817d19718efe9a6f22396fd3355fcb276db32fb
    Merge: 50b7e8b 4f999d6
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 14:56:18 2022 -0400

        Merge branch 'issue-874b' of https://github.com/StormSurgeLive/asgs into issue-874b

    commit 50b7e8b2e416bc6ee9c06cc3fdbf01aafedf55bc
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Sun Jul 10 14:56:14 2022 -0400

        Added checks for directory ownership

    commit 4f999d612b55161dfe40d29f764f5dfdf64fb9c1
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Fri Jul 8 17:02:55 2022 -0400

        Added checks for ownership and existence

    commit 19c8f4c849e8ddb9654ab8c0da768b34965e6db5
    Author: Jason Fleming <jason.fleming@seahorsecoastal.com>
    Date:   Fri Jul 8 17:02:55 2022 -0400

        Added checks for ownership and existence

Resolves #874.

(cherry picked from commit 578e199fb373bcde614d250871656964b6828c97)

+++ monitoring/logging.sh
@@ -492,16 +492,20 @@ writeASGSInstanceStatus()
 postStatus() {
     local THIS="asgs_main->monitoring/logging.sh->postStatus()"
     statfile="$statusDir/asgs.instance.status.properties"
-    logMessage "$THIS: Posting status associated with ASGS configuration and hooks in $statfile to opendap."
-    # redact username from log file and rename with .txt extension so
-    # we can review it directly in a web browser
-    local logfile=$(basename -- $SYSLOG)
-    textlog="${logfile%.*}.txt"
-    sed "s/$USER/\$USER/g" $SYSLOG > $statusDir/$textlog
-    # NB: OPENDAPPOST is defined in platforms.sh; also full path to
-    # $OPENDAPPOST is needed because of the way it's currently used in
-    # the list of $POSTPROCESS scripts defined in the ASGS config being used
-    $SCRIPTDIR/output/$OPENDAPPOST $statfile
+    if [[ -s $statfile ]]; then
+        logMessage "$THIS: Posting status associated with ASGS configuration and hooks in '$statfile' to opendap."
+        # redact username from log file and rename with .txt extension so
+        # we can review it directly in a web browser
+        local logfile=$(basename -- $SYSLOG)
+        textlog="${logfile%.*}.txt"
+        sed "s/$USER/\$USER/g" $SYSLOG > $statusDir/$textlog
+        # NB: OPENDAPPOST is defined in platforms.sh; also full path to
+        # $OPENDAPPOST is needed because of the way it's currently used in
+        # the list of $POSTPROCESS scripts defined in the ASGS config being used
+        $SCRIPTDIR/output/$OPENDAPPOST $statfile
+    else
+        logMessage "$THIS: Status file '$statfile' does not exist."
+    fi
 }
 
 initFileStatusMonitoring() {
